% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_export.R, R/spaceuse_kud.R
\name{calculate_area_metrics}
\alias{calculate_area_metrics}
\title{Calculate Spatial Polygon Surface Coverage Areas}
\usage{
calculate_area_metrics(sf_polys)

calculate_area_metrics(sf_polys)
}
\arguments{
\item{sf_polys}{An `sf` polygon layer.}
}
\value{
The original `sf` data frame with a numeric column named `area_km2`.

The original `sf` object with an `area_km2` column.
}
\description{
Evaluates geographic surface boundaries for calculated home range or core
utilization polygons and appends area values in square kilometers.

Evaluates polygon areas and appends area values in square kilometers.
}
