% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_export.R
\name{export_spaceuse_layers}
\alias{export_spaceuse_layers}
\title{Export Space-Use Layers}
\usage{
export_spaceuse_layers(sf_obj, filename)
}
\arguments{
\item{sf_obj}{An `sf` spatial object to export.}

\item{filename}{Character. Output filename without the `.gpkg` extension.}
}
\value{
Invisibly returns the path to the written GeoPackage file.
}
\description{
Exports an `sf` spatial object as a GeoPackage file.
}
