% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_kud.R
\name{get_home_range}
\alias{get_home_range}
\title{Extract Home-Range Polygons}
\usage{
get_home_range(kud, level = 95)
}
\arguments{
\item{kud}{An `EstUDm` object produced by `estimate_kernel_ud()`.}

\item{level}{Numeric. Home-range isopleth percentage. Default is `95`.}
}
\value{
An `sf` polygon object with `level_pct` and `area_type` columns.
}
\description{
Extracts the home-range utilization distribution contour, usually the 95%
isopleth.
}
