% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_export.R
\name{get_spatial_centroids}
\alias{get_spatial_centroids}
\title{Extract Spatial Centroids from Area Features}
\usage{
get_spatial_centroids(sf_polys)
}
\arguments{
\item{sf_polys}{An `sf` spatial feature collection containing target polygons.}
}
\value{
An `sf` spatial point object representing feature centroids.
}
\description{
Computes geographic centroids for calculated boundary polygons or utilization
contours.
}
