% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_overlap.R
\name{join_tracks_to_fishing_grid}
\alias{join_tracks_to_fishing_grid}
\title{Join track locations to a fisheries grid or polygon layer}
\usage{
join_tracks_to_fishing_grid(
  track_data,
  fisheries_data,
  join_type = c("intersects", "within", "nearest")
)
}
\arguments{
\item{track_data}{An sf object of bird track points or polygons.}

\item{fisheries_data}{An sf object of fisheries cells or polygons.}

\item{join_type}{Character string. One of "intersects", "within", or "nearest".}
}
\value{
An sf object containing joined track-fisheries data.
}
\description{
Spatially matches bird locations or utilization data to fisheries cells
or polygons.
}
