% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_core_use_hotspots}
\alias{plot_core_use_hotspots}
\title{Plot Home Range and Core Space-Use Hotspots}
\usage{
plot_core_use_hotspots(
  sf_polys,
  fill_col = NULL,
  title = "Seabird Space-Use Hotspots"
)
}
\arguments{
\item{sf_polys}{An `sf` polygon dataset generated by space-use functions.}

\item{fill_col}{Character. Column used to fill polygons.}

\item{title}{Character. Plot title.}
}
\value{
A `ggplot2` plot object.
}
\description{
Plot Home Range and Core Space-Use Hotspots
}
