% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_density_map}
\alias{plot_density_map}
\title{Plot Spatial Density of Seabird Use}
\usage{
plot_density_map(
  track_data,
  lon_col = "longitude",
  lat_col = "latitude",
  bins = 100,
  title = "Seabird Use Density"
)
}
\arguments{
\item{track_data}{A standardized or cleaned track data frame.}

\item{lon_col}{Character. Longitude column. Default is `"longitude"`.}

\item{lat_col}{Character. Latitude column. Default is `"latitude"`.}

\item{bins}{Integer. Number of 2D histogram bins per axis. Default is `100`.}

\item{title}{Character. Plot title.}
}
\value{
A `ggplot2` plot object.
}
\description{
Plot Spatial Density of Seabird Use
}
