% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_export.R
\name{plot_hotspot_map}
\alias{plot_hotspot_map}
\title{Plot Home Range and Core Space-Use Hotspots}
\usage{
plot_hotspot_map(sf_polys, title = "Seabird Space-Use Hotspots")
}
\arguments{
\item{sf_polys}{An \code{sf} polygon dataset generated by \code{get_isopleths}.}

\item{title}{Character. The map graphic header. Default is "Seabird Space-Use Hotspots".}
}
\value{
A functional \code{ggplot2} mapping plot object.
}
\description{
Generates a production-ready spatial map illustrating core home range boundaries
and utilization polygons without requiring downstream visualization dependencies.
}
