% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdiction_overlay.R
\name{print.management_layers}
\alias{print.management_layers}
\title{Print Management Layers Object}
\usage{
\method{print}{management_layers}(x, ...)
}
\arguments{
\item{x}{A management_layers object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Management Layers Object
}
