% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_Functions.R
\name{process_bird_data}
\alias{process_bird_data}
\title{Bird Filtering Pipeline}
\usage{
process_bird_data(data, start_dates_vec, id_col)
}
\arguments{
\item{data}{Input dataframe (expects Tag, datetime, type, correct_step_distance)}

\item{start_dates_vec}{A character vector of fledge dates}

\item{id_col}{The name of your ID column (unquoted)}
}
\value{
A filtered dataframe excluding irrelevant dates, calculated velocities, equinoxes, extreme latitudes, and land-proximate points
}
\description{
Bird Filtering Pipeline
}
