% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceuse_kud.R
\name{calculate_kud}
\alias{calculate_kud}
\title{Calculate Kernel Utilization Distributions}
\usage{
calculate_kud(tracks, ref = "href")
}
\arguments{
\item{tracks}{An `sf` point object containing GPS locations.}

\item{ref}{Bandwidth selection method or value.}
}
\value{
An `EstUDm` object.
}
\description{
Backward-compatible wrapper for `estimate_kernel_ud()`.
}
